% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data From Server}
\usage{
get_data(
  consort = "EBOPLUS",
  cohort = "usa",
  dttype = "pheno",
  dtset = "rnaseq",
  dtlevel = "gene_expression",
  url = "http://45.79.5.6:3000/raw"
)
}
\arguments{
\item{url}{}
}
\value{

}
\description{
Get Data From Server
}
