% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFData.R
\name{makeFData}
\alias{makeFData}
\title{Create Feature Data
Gene Expression - RNASeq
Gene Metadata}
\usage{
makeFData(ds, lncRNAs = NULL)
}
\arguments{
\item{lncRNAs}{}
}
\value{

}
\description{
Create Feature Data
Gene Expression - RNASeq
Gene Metadata
}
