% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_to_baseline.R
\name{norm_to_baseline}
\alias{norm_to_baseline}
\title{Normalize a Subjects Samples to its Baseline Sample
Gene Expression - RNASeq}
\usage{
norm_to_baseline(
  ds,
  tp_col = "sampling_day",
  sample_col = "sample_id",
  vol_col = "volunteer_id"
)
}
\arguments{
\item{vol_col}{}
}
\value{

}
\description{
Normalize a Subjects Samples to its Baseline Sample
Gene Expression - RNASeq
}
