% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCountsData.R
\name{normCountsData}
\alias{normCountsData}
\title{Normalize Counts Data
Normalization: CPM}
\usage{
normCountsData(ds, filter_genes = T)
}
\arguments{
\item{filter_genes}{}
}
\value{

}
\description{
Normalize Counts Data
Normalization: CPM
}
