% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_all.R
\name{preprocess_all}
\alias{preprocess_all}
\title{Preprocess All RNASeq Data Sets
Gene Expression - RNASeq}
\usage{
preprocess_all(
  counts_usa = counts_usa,
  pheno_usa = pheno_usa,
  counts_geneva = counts_geneva,
  pheno_geneva = pheno_geneva
)
}
\arguments{
\item{pheno_geneva}{}
}
\value{

}
\description{
Preprocess All RNASeq Data Sets
Gene Expression - RNASeq
}
