% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_counts_geneva.R
\name{preprocess_counts_geneva}
\alias{preprocess_counts_geneva}
\title{Preprocessing Geneva Counts Data
Gene Expression - RNASeq}
\usage{
preprocess_counts_geneva(counts)
}
\arguments{
\item{counts}{}
}
\value{

}
\description{
Preprocessing Geneva Counts Data
Gene Expression - RNASeq
}
