% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_counts_germany.R
\name{preprocess_counts_germany}
\alias{preprocess_counts_germany}
\title{Preprocess Germany Counts Data
Gene Expression - RNASeq}
\usage{
preprocess_counts_germany(counts)
}
\arguments{
\item{counts}{}
}
\value{

}
\description{
Preprocess Germany Counts Data
Gene Expression - RNASeq
}
