% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_dataset.R
\name{preprocess_dataset}
\alias{preprocess_dataset}
\title{Preprocess RNASeq Datasets}
\usage{
preprocess_dataset(counts, pheno, cohort = NULL)
}
\arguments{
\item{cohort}{}
}
\value{

}
\description{
Preprocess RNASeq Datasets
}
