% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_geneva.R
\name{preprocess_geneva}
\alias{preprocess_geneva}
\title{Preprocess Geneva dataset
(wrapper)}
\usage{
preprocess_geneva(counts = counts, pheno = pheno)
}
\arguments{
\item{pheno}{}
}
\value{

}
\description{
Preprocess Geneva dataset
(wrapper)
}
