% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_pheno_geneva.R
\name{preprocess_pheno_geneva}
\alias{preprocess_pheno_geneva}
\title{Preprocess Geneva Pheno Data}
\usage{
preprocess_pheno_geneva(pheno, print_schema = F)
}
\arguments{
\item{print_schema}{}
}
\value{

}
\description{
Preprocess Geneva Pheno Data
}
