% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_pheno_germany.R
\name{preprocess_pheno_germany}
\alias{preprocess_pheno_germany}
\title{Preprocess Germany Pheno Data
Gene Expression - RNASeq}
\usage{
preprocess_pheno_germany(counts)
}
\arguments{
\item{counts}{}
}
\value{

}
\description{
Preprocess Germany Pheno Data
Gene Expression - RNASeq
}
