% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Save Data}
\usage{
save_data(
  data,
  dttype = NULL,
  consort = NULL,
  cohort = NULL,
  dtset = "rnaseq",
  dtlevel = "gene_expression"
)
}
\arguments{
\item{dtlevel}{}
}
\value{

}
\description{
Save Data
}
